+++ src/main/java/fitpet_be/domain/repository/EstimateRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.Query;
 public interface EstimateRepository extends JpaEstimateRepository {
 
     @Query("SELECT COUNT(e) FROM Estimate e WHERE DATE(e.createdAt) = :today")
-    Long countEstimateByCreatedAtToday(@Param("today") LocalDate today);
+    Integer countEstimateByCreatedAtToday(@Param("today") LocalDate today);
 
 }
